I created another custom hook, for the nabvar component
