+++ src/main.jsx
@@ -1,13 +1,10 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import AuthContext from "./Context/AuthContext.jsx";
 import "./index.css";
 import App from "./App.jsx";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
-    <AuthContext>
-      <App />
-    </AuthContext>
+    <App />
   </StrictMode>
 );
